fix: accountType을 Enum으로 변경
