+++ data/src/main/kotlin/land/sungbin/androidprojecttemplate/data/model/AccountInformationData.kt
@@ -3,6 +3,7 @@ package land.sungbin.androidprojecttemplate.data.model
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties
+import land.sungbin.androidprojecttemplate.domain.constants.AccountType
 
 @Entity
 data class AccountInformationData(
