+++ domain/src/main/kotlin/land/sungbin/androidprojecttemplate/domain/model/AccountInformationEntity.kt
@@ -1,6 +1,8 @@
 package land.sungbin.androidprojecttemplate.domain.model
 
+import land.sungbin.androidprojecttemplate.domain.constants.AccountType
+
 data class AccountInformationEntity(
-    val accountType: String,
+    val accountType: AccountType,
     val email: String,
 )
